Split the Tx modules into top level modules
