+++ cooked-validators/src/Cooked/Attack/AddToken.hs
@@ -3,7 +3,7 @@ module Cooked.Attack.AddToken where
 import Control.Monad
 import Cooked.Attack.Tweak
 import Cooked.MockChain.Wallet
-import Cooked.Tx.Constraints.Type
+import Cooked.Skeleton
 import qualified Data.Map as Map
 import qualified Ledger as Pl
 import qualified Ledger.Value as Pl
