+++ cooked-validators/src/Cooked/Attack/DoubleSat.hs
@@ -10,7 +10,8 @@ import Control.Monad
 import Cooked.Attack.Tweak
 import Cooked.MockChain.Monad
 import Cooked.MockChain.Wallet
-import Cooked.Tx.Constraints.Type
+import Cooked.Output
+import Cooked.Skeleton
 import Data.List
 import Data.Map (Map)
 import qualified Data.Map as Map
