+++ cooked-validators/src/Cooked/Attack/DupToken.hs
@@ -5,7 +5,7 @@ module Cooked.Attack.DupToken where
 import Control.Monad
 import Cooked.Attack.Tweak
 import Cooked.MockChain.Wallet
-import Cooked.Tx.Constraints.Type
+import Cooked.Skeleton
 import qualified Ledger as L
 import qualified Ledger.Scripts as Pl
 import qualified Ledger.Value as Pl
