+++ cooked-validators/src/Cooked/Attack/Tweak/Common.hs
@@ -11,7 +11,7 @@ module Cooked.Attack.Tweak.Common where
 import Control.Monad
 import Control.Monad.State
 import Cooked.MockChain.Monad
-import Cooked.Tx.Constraints.Type
+import Cooked.Skeleton
 import Data.List
 import Data.Maybe
 import qualified Data.Set as Set
