+++ cooked-validators/src/Cooked/Attack/Tweak/OutPermutations.hs
@@ -2,7 +2,7 @@ module Cooked.Attack.Tweak.OutPermutations where
 
 import Control.Monad
 import Cooked.Attack.Tweak.Common
-import Cooked.Tx.Constraints.Type
+import Cooked.Skeleton
 
 data PermutOutTweakMode = KeepIdentity (Maybe Int) | OmitIdentity (Maybe Int)
