+++ cooked-validators/src/Cooked/MockChain/Monad/Direct.hs
@@ -25,7 +25,8 @@ import Cooked.MockChain.Monad
 import Cooked.MockChain.Monad.GenerateTx
 import Cooked.MockChain.UtxoState
 import Cooked.MockChain.Wallet
-import Cooked.Tx.Constraints.Type
+import Cooked.Output
+import Cooked.Skeleton
 import Data.Default
 import Data.Function (on)
 import Data.List
