+++ cooked-validators/src/Cooked/MockChain/Monad/GenerateTx.hs
@@ -8,7 +8,8 @@ import qualified Cardano.Api as C
 import qualified Cardano.Api.Shelley as C
 import Control.Arrow
 import Cooked.MockChain.Wallet
-import Cooked.Tx.Constraints.Type
+import Cooked.Output
+import Cooked.Skeleton
 import Data.Bifunctor
 import Data.Default
 import qualified Data.List.NonEmpty as NEList
