+++ cooked-validators/tests/Cooked/Attack/DatumHijackingSpec.hs
@@ -12,7 +12,8 @@ module Cooked.Attack.DatumHijackingSpec (tests) where
 import Control.Monad
 import Cooked.Attack
 import Cooked.MockChain
-import Cooked.Tx.Constraints.Type
+import Cooked.Output
+import Cooked.Skeleton
 import Data.Default
 import qualified Data.Map as Map
 import qualified Data.Set as Set
