+++ cooked-validators/tests/Cooked/Attack/DupTokenSpec.hs
@@ -10,8 +10,7 @@ import Cooked.Attack
 import Cooked.Currencies
 import Cooked.Ltl
 import Cooked.MockChain
-import Cooked.Tx.Constraints.Optics
-import Cooked.Tx.Constraints.Type
+import Cooked.Skeleton
 import Data.Default
 import qualified Data.Set as Set
 import qualified Ledger.Ada as L
