+++ cooked-validators/tests/Cooked/Attack/Tweak/CommonSpec.hs
@@ -3,8 +3,7 @@ module Cooked.Attack.Tweak.CommonSpec (tests) where
 import Cooked.Attack.Tweak.Common
 import Cooked.MockChain
 import Cooked.MockChain.Testing
-import Cooked.Tx.Constraints.Optics
-import Cooked.Tx.Constraints.Type
+import Cooked.Skeleton
 import Data.Default
 import qualified Ledger.Ada as L
 import qualified Ledger.Value as L
