+++ cooked-validators/tests/Cooked/InlineDatumsSpec.hs
@@ -14,7 +14,7 @@ module Cooked.InlineDatumsSpec where
 import Control.Arrow
 import Control.Monad
 import Cooked
-import Cooked.Tx.Constraints.Type
+import Cooked.Skeleton
 import Data.Default
 import qualified Data.Map as Map
 import Data.Maybe
